+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added the possibility to specify the time of a ``Record`` with ``push_back``.
+
 ## [0.1.0] - 2021-04-02
 
 First release of `yarp-telemetry`, compatible with YARP 3.4.